Changes to support int8 weight and fp32 bias in QNNPACK similar to FBGEMM.
Summary
QNNPACK still expects uint8 weights so we convert from int8 to uint8 in the operator code.
Add support for FP32 bias - Re-quantize bias during time time based on input scale.
If the value of input scale changes in the packed struct we requantize the bias with the updated input scale.

Test Plan:
    python test/test_quantized.py TestQNNPackOps

ghstack-source-id: 76c0b4faf41c56b528d51b679ba94edc9bddbd87
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26307